+++ services/question.go
@@ -22,13 +22,21 @@ type QuestionService struct {
 	QuestionRepo repositories.QuestionRepository
 }
 
-func (s *QuestionService) CreateQuestion (ctx context.Context, quizID string, body dto.CreateQuestionDTO) (entities.QuestionAPI, error){
+func (s *QuestionService) CreateQuestion (ctx context.Context, quizID string, body dto.CreateQuestionDTO, userID int) (entities.QuestionAPI, error){
 	if len(body.Answers) == 0 {
 		return entities.QuestionAPI{}, ErrNoQuestionAnswers
 	}
 	qID, err := strconv.Atoi(quizID)
 	if err != nil {
 		return entities.QuestionAPI{}, ErrInvalidIDFormat
+	}
+
+	isOwner, err := s.QuestionRepo.CheckIfQuizOwner(ctx, qID, userID)
+	if err != nil {
+		return entities.QuestionAPI{}, err
+	}
+	if !isOwner {
+		return entities.QuestionAPI{}, ErrNotAnAuthor
 	}
 
 	question, err := s.QuestionRepo.CreateQuestion(ctx, qID, body)
@@ -38,6 +46,7 @@ func (s *QuestionService) CreateQuestion (ctx context.Context, quizID string, bo
 		}
 		return entities.QuestionAPI{}, err
 	}
+
 	return question, nil
 }
 
@@ -71,13 +80,22 @@ func (s *QuestionService) GetQuestion(ctx context.Context, questionID string)(en
 	return question, nil
 }
 
-func (s *QuestionService) UpdateQuestion(ctx context.Context, questionID string, data dto.UpdateQuestionDTO) (entities.QuestionAPI ,error){
+func (s *QuestionService) UpdateQuestion(ctx context.Context, questionID string, data dto.UpdateQuestionDTO, userID int) (entities.QuestionAPI ,error){
 	qID, err := strconv.Atoi(questionID)
 	if err != nil {
 		return entities.QuestionAPI{}, ErrInvalidIDFormat
 	}
 
+	isOwner, err := s.QuestionRepo.CheckIfQuestionOwner(ctx, qID, userID)
+	if err != nil {
+		return entities.QuestionAPI{}, err
+	}
+	if !isOwner {
+		return entities.QuestionAPI{}, ErrNotAnAuthor
+	}
+
 	question, err := s.QuestionRepo.UpdateQuestion(ctx, qID, data)
+
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) || errors.Is(err, sql.ErrNoRows) {
 			return entities.QuestionAPI{}, ErrQuestionNotFound
@@ -87,13 +105,22 @@ func (s *QuestionService) UpdateQuestion(ctx context.Context, questionID string,
 	return question, nil
 }
 
-func (s *QuestionService) DeleteQuestion(ctx context.Context, questionID string) (error){
+func (s *QuestionService) DeleteQuestion(ctx context.Context, questionID string, userID int) (error){
 	qID, err := strconv.Atoi(questionID)
 	if err != nil {
 		return ErrInvalidIDFormat
 	}
 
+	isOwner, err := s.QuestionRepo.CheckIfQuestionOwner(ctx, qID, userID)
+	if err != nil {
+		return err
+	}
+	if !isOwner {
+		return ErrNotAnAuthor
+	}
+
 	err = s.QuestionRepo.DeleteQuestion(ctx, qID)
+
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) || errors.Is(err, sql.ErrNoRows) {
 			return ErrQuestionNotFound
